refactor(api-python-sdk): introduce error handling
