+++ api/python/airthings_sdk/__init__.py
@@ -1,9 +1,14 @@
 """ A client library for accessing Airthings for Consumer API """
 
-from .parser import Airthings, AirthingsDevice, AirthingsSensor
+from .mapper import Airthings
+from .types import AirthingsDevice, AirthingsSensor
+from .errors import UnexpectedStatusError, ApiError, UnexpectedPayloadError
 
 __all__ = (
     "Airthings",
     "AirthingsDevice",
     "AirthingsSensor",
+    "UnexpectedStatusError",
+    "ApiError",
+    "UnexpectedPayloadError",
 )
