+++ api/python/airthings_sdk/const.py
@@ -1,0 +1,4 @@
+"""Module providing constants for the Airthings API SDK."""
+
+AUTH_URL = "https://accounts-api.airthings.com"
+API_URL = "https://consumer-api.airthings.com"
