+++ api/python/examples/fetch_devices_and_sensors.py
@@ -1,3 +1,5 @@
+""" Example of fetching devices and sensors from Airthings API. """
+
 import sys
 from airthings_sdk import Airthings
 
@@ -6,15 +8,13 @@ if __name__ == "__main__":
         print("Please add client id and client secret as parameters.")
         print("Usage:")
         print("python fetch_devices_and_sensors.py <client_id> <client_secret>")
-        exit(1)
+        sys.exit(1)
     client_id = sys.argv[1]
     client_secret = sys.argv[2]
 
     airthings = Airthings(
-        client_id=client_id,
-        client_secret=client_secret,
-        is_metric=True,
+        client_id=client_id, client_secret=client_secret, is_metric=True
     )
 
     devices = airthings.update_devices()
-    print(devices)
+    print(airthings.devices)
